API Manual

MadDiffModels.MadDiffModels — Module

MadDiffModels

MadDiffModels is a submodule of MadDiff. MadDiffModels allows modeling nonlinear optimization problem of the following form:

minimize:   f(x)
subject to: xl ≤   x  ≤ xu
            gl ≤ g(x) ≤ gu,

where:

x ∈ R^n is the decision variable.

f : R^n -> R is the objective function

g : R^n -> R^m is the constraint mapping.

The model is constructed as an NLPModel (see https://github.com/JuliaSmoothOptimizers/NLPModels.jl), and can be solved with various NLP solvers such as:

MadNLP (https://github.com/MadNLP/MadNLP.jl)

Ipopt (https://github.com/JuliaSmoothOptimizers/NLPModelsIpopt.jl)

Knitro (https://github.com/JuliaSmoothOptimizers/NLPModelsKnitro.jl)

MadDiffModels.Constraint — Type

Constraint

A constraint index of MadDiffModel.

MadDiffModels.MadDiffModel — Type

MadDiffModel{T <: Real}

A mathematical model of a nonlinaer program.

MadDiffModels.MadDiffModel — Method

MadDiffModel()

Creates an empty MadDiffModel{Float64}.

Example m = MadDiffModel(linear_solver = "ma27")

MadDiffModels.MadDiffModel — Method

MadDiffModel{T}()

Creates an empty MadDiffModel{T}.

Example m = MadDiffModel{Float32}()

MadDiffModels.ModelComponent — Type

ModelComponent

A model component (eitehr a variable or a parameter) of MadDiffModel.

MadDiffModels.constraint — Method

constraint(m::MadDiffModel, e::MadDiffCore.Expression; lb=0., ub=0.)

Adds a constraint to MadDiffModel. Example ``` m = MadDiffModel()

x = [variable(m) for i=1:3] constraint(m, x[1]^2 + 2*sin(x[2]) - exp(x[3]) >= 0) constraint(m, x[1]^4+ x[2]^4 x[3]^4; lb = 0.1, ub = 1.)

MadDiffModels.dual — Method

dual(c::Constraint)

Retrun the dual of constraint c.

MadDiffModels.instantiate! — Method

instantiate!(m::MadDiffModel; sparse = true)

Instantiates the model m. The model must be instantiated before solving.

Example

using MadDiff, NLPModelsIpopt

m = MadDiffModel() 

x = [variable(m) for i=1:3]
objective(m, x[1]^2 + x[2]^2 + sin(x[3]))
constraint(m, 3x[2]^2 <= 1.)

instantiate!(m)
ipopt(m)

MadDiffModels.lower_bound — Method

lower_bound(c::Constraint)

Retrun the lower bound of constraint c.

MadDiffModels.lower_bound — Method

lower_bound(x::ModelComponent{V}) where V <: MadDiffCore.Variable

Retrun the lower bound of variable x.

MadDiffModels.objective — Method

objective(m::MadDiffModel, e::MadDiffCore.Expression

Sets the objective function for MadDiffModel. Only minimization is supported. Example

m = MadDiffModel()

x = [variable(m) for i=1:3]
objective(m, x[1]^2 + x[2] + sin(x[3]))

MadDiffModels.parameter — Method

parameter(m::MadDiffModel{T}, val)

Creates a parameter for MadDiffModel with value val. Example

m = MadDiffModel()

p = parameter(m, 0.5)

MadDiffModels.set_lower_bound — Method

set_lower_bound(c::Constraint,val)

Set the lower bound of constraint c to val.

MadDiffModels.set_lower_bound — Method

set_lower_bound(x::ModelComponent{V},val) where V <: MadDiffCore.Variable

Set the lower bound of variable 'x' to val.

MadDiffModels.set_upper_bound — Method

set_upper_bound(c::Constraint,val)

Set the upper bound of constraint c to val.

MadDiffModels.set_upper_bound — Method

set_upper_bound(x::ModelComponent{V},val) where V <: MadDiffCore.Variable

Set the upper bound of variable 'x' to val.

MadDiffModels.setvalue — Method

setvalue(p::ModelComponent{P},val) where P <: MadDiffCore.Parameter

Set the value of parameter 'p' to val.

MadDiffModels.setvalue — Method

setvalue(x::ModelComponent{V},val) where V <: MadDiffCore.Variable

Set the value of variable 'x' to val.

MadDiffModels.upper_bound — Method

upper_bound(c::Constraint)

Retrun the upper bound of constraint c.

MadDiffModels.upper_bound — Method

upper_bound(x::ModelComponent{V}) where V <: MadDiffCore.Variable

Retrun the upper bound of variable x.

MadDiffModels.value — Method

value(p::ModelComponent{P}) where P <: MadDiffCore.Parameter

Return the value of parameter p.

MadDiffModels.value — Method

value(x::ModelComponent{V}) where V <: MadDiffCore.Variable

Return the value of variable x.

MadDiffModels.variable — Method

variable(m::MadDiffModel{T}; lb=-Inf, ub=Inf, start=0.)

Creates a variable for MadDiffModel.

Example

m = MadDiffModel()

x = variable(m; lb = -1, ub = 1, start = 0.5)

MadDiffMOI.MadDiffMOI — Module

MadDiffMOI

MadDiffMOI is a submodule of MadDiff. MadDifMOI allows solving nonlinear optimization problems specified by MathOptInterface (https://github.com/jump-dev/JuMP.jl/tree/od/moi-nonlinear).

MadDiffCore.Expression — Method

Expression(ex::MOI.Nonlinear.Expression; subex = nothing)

Create a MadDiff.Expression from MOI.Expression.

MadDiffMOI.MadDiffAD — Type

MadDiffAD() <: MOI.Nonlinear.AbstractAutomaticDifferentiation

A differentiation backend for MathOptInterface based on MadDiff

MadDiffMOI.MadDiffEvaluator — Type

MadDiffEvaluator <: MOI.AbstractNLPEvaluator

A type for callbacks for MathOptInterface's nonlinear model.

MathOptInterface.Nonlinear.Evaluator — Method

MOI.Nonlinear.Evaluator(model::MOI.Nonlinear.Model, ::MadDiffAD, ::Vector{MOI.VariableIndex})

Create a MOI.Nonlinear.Evaluator from MOI.Nonlinear.Model using MadDiff's AD capability.