MadDiffModels

MadDiffModels.MadDiffModels — Module

MadDiffModels

MadDiffModels is a submodule of MadDiff. MadDiffModels allows modeling nonlinear optimization problem of the following form:

minimize:   f(x)
subject to: xl ≤   x  ≤ xu
            gl ≤ g(x) ≤ gu,

where:

x ∈ R^n is the decision variable.

f : R^n -> R is the objective function

g : R^n -> R^m is the constraint mapping.

The model is constructed as an NLPModel (see https://github.com/JuliaSmoothOptimizers/NLPModels.jl), and can be solved with various NLP solvers such as:

MadNLP (https://github.com/MadNLP/MadNLP.jl)

Ipopt (https://github.com/JuliaSmoothOptimizers/NLPModelsIpopt.jl)

Knitro (https://github.com/JuliaSmoothOptimizers/NLPModelsKnitro.jl)

MadDiffModels.Constraint — Type

Constraint

A constraint index of MadDiffModel.

MadDiffModels.MadDiffModel — Type

MadDiffModel{T <: Real}

A mathematical model of a nonlinaer program.

MadDiffModels.MadDiffModel — Method

MadDiffModel()

Creates an empty MadDiffModel{Float64}.

Example m = MadDiffModel(linear_solver = "ma27")

MadDiffModels.MadDiffModel — Method

MadDiffModel{T}()

Creates an empty MadDiffModel{T}.

Example m = MadDiffModel{Float32}()

MadDiffModels.ModelComponent — Type

ModelComponent

A model component (eitehr a variable or a parameter) of MadDiffModel.

MadDiffModels.constraint — Method

constraint(m::MadDiffModel, e::MadDiffCore.Expression; lb=0., ub=0.)

Adds a constraint to MadDiffModel. Example ``` m = MadDiffModel()

x = [variable(m) for i=1:3] constraint(m, x[1]^2 + 2*sin(x[2]) - exp(x[3]) >= 0) constraint(m, x[1]^4+ x[2]^4 x[3]^4; lb = 0.1, ub = 1.)

MadDiffModels.dual — Method

dual(c::Constraint)

Retrun the dual of constraint c.

MadDiffModels.instantiate! — Method

instantiate!(m::MadDiffModel; sparse = true)

Instantiates the model m. The model must be instantiated before solving.

Example

using MadDiff, NLPModelsIpopt

m = MadDiffModel() 

x = [variable(m) for i=1:3]
objective(m, x[1]^2 + x[2]^2 + sin(x[3]))
constraint(m, 3x[2]^2 <= 1.)

instantiate!(m)
ipopt(m)

MadDiffModels.lower_bound — Method

lower_bound(c::Constraint)

Retrun the lower bound of constraint c.

MadDiffModels.lower_bound — Method

lower_bound(x::ModelComponent{V}) where V <: MadDiffCore.Variable

Retrun the lower bound of variable x.

MadDiffModels.objective — Method

objective(m::MadDiffModel, e::MadDiffCore.Expression

Sets the objective function for MadDiffModel. Only minimization is supported. Example

m = MadDiffModel()

x = [variable(m) for i=1:3]
objective(m, x[1]^2 + x[2] + sin(x[3]))

MadDiffModels.parameter — Method

parameter(m::MadDiffModel{T}, val)

Creates a parameter for MadDiffModel with value val. Example

m = MadDiffModel()

p = parameter(m, 0.5)

MadDiffModels.set_lower_bound — Method

set_lower_bound(c::Constraint,val)

Set the lower bound of constraint c to val.

MadDiffModels.set_lower_bound — Method

set_lower_bound(x::ModelComponent{V},val) where V <: MadDiffCore.Variable

Set the lower bound of variable 'x' to val.

MadDiffModels.set_upper_bound — Method

set_upper_bound(c::Constraint,val)

Set the upper bound of constraint c to val.

MadDiffModels.set_upper_bound — Method

set_upper_bound(x::ModelComponent{V},val) where V <: MadDiffCore.Variable

Set the upper bound of variable 'x' to val.

MadDiffModels.setvalue — Method

setvalue(p::ModelComponent{P},val) where P <: MadDiffCore.Parameter

Set the value of parameter 'p' to val.

MadDiffModels.setvalue — Method

setvalue(x::ModelComponent{V},val) where V <: MadDiffCore.Variable

Set the value of variable 'x' to val.

MadDiffModels.upper_bound — Method

upper_bound(c::Constraint)

Retrun the upper bound of constraint c.

MadDiffModels.upper_bound — Method

upper_bound(x::ModelComponent{V}) where V <: MadDiffCore.Variable

Retrun the upper bound of variable x.

MadDiffModels.value — Method

value(p::ModelComponent{P}) where P <: MadDiffCore.Parameter

Return the value of parameter p.

MadDiffModels.value — Method

value(x::ModelComponent{V}) where V <: MadDiffCore.Variable

Return the value of variable x.

MadDiffModels.variable — Method

variable(m::MadDiffModel{T}; lb=-Inf, ub=Inf, start=0.)

Creates a variable for MadDiffModel.

Example

m = MadDiffModel()

x = variable(m; lb = -1, ub = 1, start = 0.5)

NLPModels.cons! — Method

NLPModels.cons!(m::MadDiffModel,x::AbstractVector,y::AbstractVector)

Evaluate the constraints of m at x and store the result in the vector y.

NLPModels.get_ifix — Function

get_ifix(m::MadDiffModel)

Return the value ifix from MadDiffModel.

NLPModels.get_ifree — Function

get_ifree(m::MadDiffModel)

Return the value ifree from MadDiffModel.

NLPModels.get_iinf — Function

get_iinf(m::MadDiffModel)

Return the value iinf from MadDiffModel.

NLPModels.get_ilow — Function

get_ilow(m::MadDiffModel)

Return the value ilow from MadDiffModel.

NLPModels.get_irng — Function

get_irng(m::MadDiffModel)

Return the value irng from MadDiffModel.

NLPModels.get_islp — Function

get_islp(m::MadDiffModel)

Return the value islp from MadDiffModel.

NLPModels.get_iupp — Function

get_iupp(m::MadDiffModel)

Return the value iupp from MadDiffModel.

NLPModels.get_jfix — Function

get_jfix(m::MadDiffModel)

Return the value jfix from MadDiffModel.

NLPModels.get_jfree — Function

get_jfree(m::MadDiffModel)

Return the value jfree from MadDiffModel.

NLPModels.get_jinf — Function

get_jinf(m::MadDiffModel)

Return the value jinf from MadDiffModel.

NLPModels.get_jlow — Function

get_jlow(m::MadDiffModel)

Return the value jlow from MadDiffModel.

NLPModels.get_jrng — Function

get_jrng(m::MadDiffModel)

Return the value jrng from MadDiffModel.

NLPModels.get_jupp — Function

get_jupp(m::MadDiffModel)

Return the value jupp from MadDiffModel.

NLPModels.get_lcon — Function

get_lcon(m::MadDiffModel)

Return the value lcon from MadDiffModel.

NLPModels.get_lin — Function

get_lin(m::MadDiffModel)

Return the value lin from MadDiffModel.

NLPModels.get_lin_nnzj — Function

get_lin_nnzj(m::MadDiffModel)

Return the value lin_nnzj from MadDiffModel.

NLPModels.get_lvar — Function

get_lvar(m::MadDiffModel)

Return the value lvar from MadDiffModel.

NLPModels.get_minimize — Function

get_minimize(m::MadDiffModel)

Return the value minimize from MadDiffModel.

NLPModels.get_name — Function

get_name(m::MadDiffModel)

Return the value name from MadDiffModel.

NLPModels.get_ncon — Function

get_ncon(m::MadDiffModel)

Return the value ncon from MadDiffModel.

NLPModels.get_nlin — Function

get_nlin(m::MadDiffModel)

Return the value nlin from MadDiffModel.

NLPModels.get_nln — Function

get_nln(m::MadDiffModel)

Return the value nln from MadDiffModel.

NLPModels.get_nln_nnzj — Function

get_nln_nnzj(m::MadDiffModel)

Return the value nln_nnzj from MadDiffModel.

NLPModels.get_nlvb — Function

get_nlvb(m::MadDiffModel)

Return the value nlvb from MadDiffModel.

NLPModels.get_nlvc — Function

get_nlvc(m::MadDiffModel)

Return the value nlvc from MadDiffModel.

NLPModels.get_nlvo — Function

get_nlvo(m::MadDiffModel)

Return the value nlvo from MadDiffModel.

NLPModels.get_nnln — Function

get_nnln(m::MadDiffModel)

Return the value nnln from MadDiffModel.

NLPModels.get_nnzh — Function

get_nnzh(m::MadDiffModel)

Return the value nnzh from MadDiffModel.

NLPModels.get_nnzj — Function

get_nnzj(m::MadDiffModel)

Return the value nnzj from MadDiffModel.

NLPModels.get_nnzo — Function

get_nnzo(m::MadDiffModel)

Return the value nnzo from MadDiffModel.

NLPModels.get_nvar — Function

get_nvar(m::MadDiffModel)

Return the value nvar from MadDiffModel.

NLPModels.get_ucon — Function

get_ucon(m::MadDiffModel)

Return the value ucon from MadDiffModel.

NLPModels.get_uvar — Function

get_uvar(m::MadDiffModel)

Return the value uvar from MadDiffModel.

NLPModels.get_x0 — Function

get_x0(m::MadDiffModel)

Return the value x0 from MadDiffModel.

NLPModels.get_y0 — Function

get_y0(m::MadDiffModel)

Return the value y0 from MadDiffModel.

NLPModels.grad! — Method

NLPModels.grad!(m::MadDiffModel,x::AbstractVector,y::AbstractVector)

Evaluate the gradient of m at x and store the result in the vector y.

NLPModels.hess_coord! — Method

NLPModels.hess_coord!(m::MadDiffModel,x::AbstractVector,lag::AbstractVector,z::AbstractVector; obj_weight = 1.0)

Evaluate the Lagrangian Hessian of m at primal x, dual lag, and objective weight obj_weight and store the result in the vector zin sparse coordinate format.

NLPModels.hess_structure! — Method

NLPModels.hess_structure!(m::MadDiffModel,I::AbstractVector{T},J::AbstractVector{T})

Evaluate the structure of the Lagrangian Hessian and store the result in I and J in sparse coordinate format.

NLPModels.jac_coord! — Method

NLPModels.jac_coord!(m::MadDiffModel,x::AbstractVector,J::AbstractVector)

Evaluate the constraints Jacobian of m at x and store the result in the vector J in sparse coordinate format.

NLPModels.jac_structure! — Method

NLPModels.jac_structure!(m::MadDiffModel,I::AbstractVector{T},J::AbstractVector{T})

Evaluate the structure of the constraints Jacobian and store the result in I and J in sparse coordinate format.

NLPModels.obj — Method

NLPModels.obj(m::MadDiffModel,x::AbstractVector)

Return the objective value of m at x.