MadDiffMOI

MadDiffMOI.MadDiffMOI — Module

MadDiffMOI

MadDiffMOI is a submodule of MadDiff. MadDifMOI allows solving nonlinear optimization problems specified by MathOptInterface (https://github.com/jump-dev/JuMP.jl/tree/od/moi-nonlinear).

MadDiffCore.Expression — Method

Expression(ex::MOI.Nonlinear.Expression; subex = nothing)

Create a MadDiff.Expression from MOI.Expression.

MadDiffCore.SparseNLPCore — Method

MadDiffCore.SparseNLPCore(nlp_data::MOI.Nonlinear.Model)

Create MadDiffCore.SparseNLPCore from MOI.Nonlinear.Model.

MadDiffMOI.MadDiffAD — Type

MadDiffAD() <: MOI.Nonlinear.AbstractAutomaticDifferentiation

A differentiation backend for MathOptInterface based on MadDiff

MadDiffMOI.MadDiffEvaluator — Type

MadDiffEvaluator <: MOI.AbstractNLPEvaluator

A type for callbacks for MathOptInterface's nonlinear model.

MathOptInterface.NLPBlockData — Method

MOI.NLPBlockData(evaluator::MadDiffEvaluator)

Create MOI.NLPBlockData from MadDiffEvaluator

MathOptInterface.Nonlinear.Evaluator — Method

MOI.Nonlinear.Evaluator(model::MOI.Nonlinear.Model, ::MadDiffAD, ::Vector{MOI.VariableIndex})

Create a MOI.Nonlinear.Evaluator from MOI.Nonlinear.Model using MadDiff's AD capability.

MathOptInterface.eval_constraint — Method

MOI.eval_constraint(evaluator::MadDiffEvaluator, g, x)

Evaluate the gradient of evaluator at x and store the result in the vector g.

MathOptInterface.eval_constraint_jacobian — Method

MOI.eval_constraint_jacobian(evaluator::MadDiffEvaluator, J, x)

Evaluate the constraints Jacobian of evaluator at x and store the result in the vector J in sparse coordinate format.

MathOptInterface.eval_hessian_lagrangian — Method

MOI.eval_hessian_lagrangian(evaluator::MadDiffEvaluator, H, x, σ, μ)

Evaluate the Lagrangian Hessian of evaluator at primal x, dual μ, and objective weight σ and store the result in the vector H in sparse coordinate format.

MathOptInterface.eval_objective — Method

MOI.eval_objective(evaluator::MadDiffEvaluator, x)

Return the objective value of evaluator at x.

MathOptInterface.eval_objective_gradient — Method

MOI.eval_objective_gradient(evaluator::MadDiffEvaluator, g, x)

Evaluate the constraints of evaluator at x and store the result in the vector g.

MathOptInterface.hessian_lagrangian_structure — Method

MOI.hessian_lagrangian_structure(evaluator::MadDiffEvaluator)

Return the structure of the Lagrangian Hessian in Vector{Tuple{Int,Int}} format.

MathOptInterface.jacobian_structure — Method

MOI.jacobian_structure(evaluator::MadDiffEvaluator)

Return the structure of the constraints Jacobian in Vector{Tuple{Int,Int}} format.